Struct sc_rpc_api::chain::gen_client::Client

pub struct Client<Number, Hash, Header, SignedBlock> { /* fields omitted */ }

The Client.

Implementations

impl<Number, Hash, Header, SignedBlock> Client<Number, Hash, Header, SignedBlock> where
Number: Send + Sync + 'static,
Hash: Send + Sync + 'static + DeserializeOwned + Serialize,
Header: Send + Sync + 'static + DeserializeOwned,
SignedBlock: Send + Sync + 'static + DeserializeOwned,

pub fn new(sender: RpcChannel) -> Self

Creates a new Client.

pub fn header(
&self,
hash: Option<Hash>
) -> impl Future<Item = Option<Header>, Error = RpcError>

Get header of a relay chain block.

pub fn block(
&self,
hash: Option<Hash>
) -> impl Future<Item = Option<SignedBlock>, Error = RpcError>

Get header and body of a relay chain block.

pub fn block_hash(
&self,
hash: Option<ListOrValue<NumberOrHex>>
) -> impl Future<Item = ListOrValue<Option<Hash>>, Error = RpcError>

Get hash of the n-th block in the canon chain.

By default returns latest block hash.

pub fn finalized_head(&self) -> impl Future<Item = Hash, Error = RpcError>

Get hash of the last finalized block in the canon chain.

pub fn subscribe_all_heads(
&self
) -> impl Future<Item = TypedSubscriptionStream<Header>, Error = RpcError>

All head subscription

pub fn subscribe_new_heads(
&self
) -> impl Future<Item = TypedSubscriptionStream<Header>, Error = RpcError>

New head subscription

pub fn subscribe_finalized_heads(
&self
) -> impl Future<Item = TypedSubscriptionStream<Header>, Error = RpcError>

Finalized head subscription