Trait sc_rpc_api::child_state::ChildStateApiServer

pub trait ChildStateApiServer<Hash>: Sized + Send + Sync + 'static {
    fn storage_keys<'life0, 'async_trait>(
        &'life0 self, 
        child_storage_key: PrefixedStorageKey, 
        prefix: StorageKey, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<StorageKey>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn storage_keys_paged<'life0, 'async_trait>(
        &'life0 self, 
        child_storage_key: PrefixedStorageKey, 
        prefix: Option<StorageKey>, 
        count: u32, 
        start_key: Option<StorageKey>, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<StorageKey>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn storage<'life0, 'async_trait>(
        &'life0 self, 
        child_storage_key: PrefixedStorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<StorageData>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn storage_entries<'life0, 'async_trait>(
        &'life0 self, 
        child_storage_key: PrefixedStorageKey, 
        keys: Vec<StorageKey>, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Option<StorageData>>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn storage_hash<'life0, 'async_trait>(
        &'life0 self, 
        child_storage_key: PrefixedStorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Hash>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn storage_size<'life0, 'async_trait>(
        &'life0 self, 
        child_storage_key: PrefixedStorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<u64>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;
    fn read_child_proof<'life0, 'async_trait>(
        &'life0 self, 
        child_storage_key: PrefixedStorageKey, 
        keys: Vec<StorageKey>, 
        hash: Option<Hash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<ReadProof<Hash>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn into_rpc(self) -> RpcModule<Self>
    where
        Hash: Send + Sync + 'static + DeserializeOwned + Serialize,
    { ... }
}

Server trait implementation for the ChildStateApi RPC API.

Required methods

fn storage_keys<'life0, 'async_trait>(
&'life0 self,
child_storage_key: PrefixedStorageKey,
prefix: StorageKey,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<StorageKey>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns the keys with prefix from a child storage, leave empty to get all the keys

fn storage_keys_paged<'life0, 'async_trait>(
&'life0 self,
child_storage_key: PrefixedStorageKey,
prefix: Option<StorageKey>,
count: u32,
start_key: Option<StorageKey>,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<StorageKey>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns the keys with prefix from a child storage with pagination support. Up to count keys will be returned. If start_key is passed, return next keys in storage in lexicographic order.

fn storage<'life0, 'async_trait>(
&'life0 self,
child_storage_key: PrefixedStorageKey,
key: StorageKey,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = RpcResult<Option<StorageData>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns a child storage entry at a specific block’s state.

fn storage_entries<'life0, 'async_trait>(
&'life0 self,
child_storage_key: PrefixedStorageKey,
keys: Vec<StorageKey>,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Option<StorageData>>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns child storage entries for multiple keys at a specific block’s state.

fn storage_hash<'life0, 'async_trait>(
&'life0 self,
child_storage_key: PrefixedStorageKey,
key: StorageKey,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = RpcResult<Option<Hash>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns the hash of a child storage entry at a block’s state.

fn storage_size<'life0, 'async_trait>(
&'life0 self,
child_storage_key: PrefixedStorageKey,
key: StorageKey,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = RpcResult<Option<u64>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns the size of a child storage entry at a block’s state.

fn read_child_proof<'life0, 'async_trait>(
&'life0 self,
child_storage_key: PrefixedStorageKey,
keys: Vec<StorageKey>,
hash: Option<Hash>
) -> Pin<Box<dyn Future<Output = RpcResult<ReadProof<Hash>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Returns proof of storage for child key entries at a specific block’s state.

Provided methods

fn into_rpc(self) -> RpcModule<Self> where
Hash: Send + Sync + 'static + DeserializeOwned + Serialize,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.