Trait sc_rpc_api::dev::DevApiClient

pub trait DevApiClient<Hash>: ClientTwhere
    Hash: Send + Sync + 'static + Serialize,{
    fn block_stats<'life0, 'async_trait>(
        &'life0 self,
        block_hash: Hash
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<BlockStats>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Client implementation for the DevApi RPC API.

Provided Methods

fn block_stats<'life0, 'async_trait>(
&'life0 self,
block_hash: Hash
) -> Pin<Box<dyn Future<Output = RpcResult<Option<BlockStats>>> + Send + 'async_trait>>where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Reexecute the specified block_hash and gather statistics while doing so.

This function requires the specified block and its parent to be available at the queried node. If either the specified block or the parent is pruned, this function will return None.