Trait sc_rpc_api::dev::DevApiServer

pub trait DevApiServer<Hash>: Sized + Send + Sync + 'static {
    fn block_stats(&self, block_hash: Hash) -> RpcResult<Option<BlockStats>>;

    fn into_rpc(self) -> RpcModule<Self>
    where
        Hash: Send + Sync + 'static + DeserializeOwned,
    { ... }
}

Server trait implementation for the DevApi RPC API.

Required Methods

fn block_stats(&self, block_hash: Hash) -> RpcResult<Option<BlockStats>>

Reexecute the specified block_hash and gather statistics while doing so.

This function requires the specified block and its parent to be available at the queried node. If either the specified block or the parent is pruned, this function will return None.

Provided Methods

fn into_rpc(self) -> RpcModule<Self>where
Hash: Send + Sync + 'static + DeserializeOwned,

Collects all the methods and subscriptions defined in the trait and adds them into a single RpcModule.