Trait sc_rpc_api::offchain::OffchainApiClient

pub trait OffchainApiClient: ClientT {
    fn set_local_storage<'life0, 'async_trait>(
        &'life0 self, 
        kind: StorageKind, 
        key: Bytes, 
        value: Bytes
    ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
    fn get_local_storage<'life0, 'async_trait>(
        &'life0 self, 
        kind: StorageKind, 
        key: Bytes
    ) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: Sync + 'async_trait,
    { ... }
}

Client implementation for the OffchainApi RPC API.

Provided methods

fn set_local_storage<'life0, 'async_trait>(
&'life0 self,
kind: StorageKind,
key: Bytes,
value: Bytes
) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Set offchain local storage under given key and prefix.

fn get_local_storage<'life0, 'async_trait>(
&'life0 self,
kind: StorageKind,
key: Bytes
) -> Pin<Box<dyn Future<Output = RpcResult<Option<Bytes>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: Sync + 'async_trait,

Get offchain local storage under given key and prefix.