Trait sc_rpc_api::statement::StatementApiClient

pub trait StatementApiClient: ClientT {
    fn dump<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn broadcasts<'life0, 'async_trait>(
        &'life0 self,
        match_all_topics: Vec<[u8; 32]>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn posted<'life0, 'async_trait>(
        &'life0 self,
        match_all_topics: Vec<[u8; 32]>,
        dest: [u8; 32]
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn posted_clear<'life0, 'async_trait>(
        &'life0 self,
        match_all_topics: Vec<[u8; 32]>,
        dest: [u8; 32]
    ) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn submit<'life0, 'async_trait>(
        &'life0 self,
        encoded: Bytes
    ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn remove<'life0, 'async_trait>(
        &'life0 self,
        statement_hash: [u8; 32]
    ) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
}

Client implementation for the StatementApi RPC API.

Provided Methods

fn dump<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Return all statements, SCALE-encoded.

fn broadcasts<'life0, 'async_trait>(
&'life0 self,
match_all_topics: Vec<[u8; 32]>
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Return the data of all known statements which include all topics and have no DecryptionKey field.

fn posted<'life0, 'async_trait>(
&'life0 self,
match_all_topics: Vec<[u8; 32]>,
dest: [u8; 32]
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Return the data of all known statements whose decryption key is identified as dest (this will generally be the public key or a hash thereof for symmetric ciphers, or a hash of the private key for symmetric ciphers).

fn posted_clear<'life0, 'async_trait>(
&'life0 self,
match_all_topics: Vec<[u8; 32]>,
dest: [u8; 32]
) -> Pin<Box<dyn Future<Output = RpcResult<Vec<Bytes>>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Return the decrypted data of all known statements whose decryption key is identified as dest. The key must be available to the client.

fn submit<'life0, 'async_trait>(
&'life0 self,
encoded: Bytes
) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Submit a pre-encoded statement.

fn remove<'life0, 'async_trait>(
&'life0 self,
statement_hash: [u8; 32]
) -> Pin<Box<dyn Future<Output = RpcResult<()>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Remove a statement from the store.