Crate sc_rpc_api

Substrate RPC interfaces.

A collection of RPC methods and subscriptions supported by all substrate clients.

Modules

author

Substrate block-author/full-node API.

chain

Substrate blockchain API.

child_state

Substrate child state API

dev

Substrate dev API containing RPCs that are mainly meant for debugging and stats collection for developers. The endpoints in this RPC module are not meant to be available to non-local users and are all marked unsafe.

mixnet

Substrate mixnet API.

offchain

Substrate offchain API.

state

Substrate state API.

statement

Substrate Statement Store RPC API.

system

Substrate system API.

Enums

DenyUnsafe

Signifies whether a potentially unsafe RPC should be denied.