Trait sc_rpc_api::child_state::ChildStateApi

pub trait ChildStateApi<Hash>: Sized + Send + Sync + 'static {
    type Metadata: Metadata;
    fn storage_keys(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        prefix: StorageKey, 
        hash: Option<Hash>
    ) -> FutureResult<Vec<StorageKey>>;

    fn storage_keys_paged(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        prefix: Option<StorageKey>, 
        count: u32, 
        start_key: Option<StorageKey>, 
        hash: Option<Hash>
    ) -> FutureResult<Vec<StorageKey>>;

    fn storage(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> FutureResult<Option<StorageData>>;

    fn storage_entries(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        keys: Vec<StorageKey>, 
        hash: Option<Hash>
    ) -> FutureResult<Vec<Option<StorageData>>>;

    fn storage_hash(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> FutureResult<Option<Hash>>;

    fn storage_size(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        key: StorageKey, 
        hash: Option<Hash>
    ) -> FutureResult<Option<u64>>;

    fn read_child_proof(
        &self, 
        child_storage_key: PrefixedStorageKey, 
        keys: Vec<StorageKey>, 
        hash: Option<Hash>
    ) -> FutureResult<ReadProof<Hash>>;

    fn to_delegate(self) -> IoDelegate<Self, Self::Metadata>
    where
        Hash: Send + Sync + 'static + Serialize + DeserializeOwned,
    { ... }
}

Substrate child state API

Note that all PrefixedStorageKey are deserialized from json and not guaranteed valid.

Associated Types

type Metadata: Metadata

RPC Metadata

Required methods

fn storage_keys(
&self,
child_storage_key: PrefixedStorageKey,
prefix: StorageKey,
hash: Option<Hash>
) -> FutureResult<Vec<StorageKey>>

DEPRECATED: Please use childstate_getKeysPaged with proper paging support. Returns the keys with prefix from a child storage, leave empty to get all the keys

fn storage_keys_paged(
&self,
child_storage_key: PrefixedStorageKey,
prefix: Option<StorageKey>,
count: u32,
start_key: Option<StorageKey>,
hash: Option<Hash>
) -> FutureResult<Vec<StorageKey>>

Returns the keys with prefix from a child storage with pagination support. Up to count keys will be returned. If start_key is passed, return next keys in storage in lexicographic order.

fn storage(
&self,
child_storage_key: PrefixedStorageKey,
key: StorageKey,
hash: Option<Hash>
) -> FutureResult<Option<StorageData>>

Returns a child storage entry at a specific block’s state.

fn storage_entries(
&self,
child_storage_key: PrefixedStorageKey,
keys: Vec<StorageKey>,
hash: Option<Hash>
) -> FutureResult<Vec<Option<StorageData>>>

Returns child storage entries for multiple keys at a specific block’s state.

fn storage_hash(
&self,
child_storage_key: PrefixedStorageKey,
key: StorageKey,
hash: Option<Hash>
) -> FutureResult<Option<Hash>>

Returns the hash of a child storage entry at a block’s state.

fn storage_size(
&self,
child_storage_key: PrefixedStorageKey,
key: StorageKey,
hash: Option<Hash>
) -> FutureResult<Option<u64>>

Returns the size of a child storage entry at a block’s state.

fn read_child_proof(
&self,
child_storage_key: PrefixedStorageKey,
keys: Vec<StorageKey>,
hash: Option<Hash>
) -> FutureResult<ReadProof<Hash>>

Returns proof of storage for child key entries at a specific block’s state.

Provided methods

fn to_delegate(self) -> IoDelegate<Self, Self::Metadata> where
Hash: Send + Sync + 'static + Serialize + DeserializeOwned,

Create an IoDelegate, wiring rpc calls to the trait methods.